Fix META.ab for toplevel usage
syntax should not be used as a predicate of requires alongside with
toploop. I don't know the meaning of "-syntax option is present on the
commandline"
(http://projects.camlcity.org/projects/dl/findlib-1.4.1/doc/ref-html/r681.html)
in this case, but this certainly does not work in the toplevel
(i.e. camlp4o does not automatically gets loaded), contrary to other
packages that does not have this problem (sexplib for example)